update health schema, start updating movement schema

+++ Assets/Scripts/Business/HealthPoints/HealthUpdatedArgs.cs
@@ -4,16 +4,16 @@ namespace Business.HealthPoints
 {
     public class HealthUpdatedArgs: EventArgs
     {
-        public HealthUpdatedArgs(int oldHealth, int newHealth, Health health)
+        public HealthUpdatedArgs(float oldHealth, float newHealth, Health health)
         {
             OldHealth = oldHealth;
             NewHealth = newHealth;
             Health = health;
         }
 
-        public int OldHealth { get; set; }
+        public float OldHealth { get; set; }
         
-        public int NewHealth { get; set; }
+        public float NewHealth { get; set; }
         
         public Health Health { get; set; }
     }
